Prevented retaining state when navigating forward
Navigate from A -> B -> C and scroll, then browser back to B then navigate to C again don't want to retain scroll position (and react state). Only want to retain scroll and state when browser forward - navigate forward should always be new scene. Tried to create different keys but can't do it - the keys are created by the NavigationStack but only the NavigationAnimation knows what the cached keys are - if go browser forward twice then how does NavigationStack know what the matching cached keys are (because need to match on browser forward)!
Instead created subkeys in NavigationAnimation - if it's not a history navigation and it matches to an unmounted scene then increment the subkey. The stack can then render the scene with that subkey to clear it!