Fixed fluently navigating forward or back multiple
Unit tested going browser forward 2 then back 1 and it failed. The back 1 lost its state. Scene B was still marked as popExit after the forward 2 so it got set to unmounted - then when the back happened it's subkey was updated and so it lost its state.
Changed so that when an unmounted is added again it gets set to pushEnter. That fixed the test.
Tried it for real to see if could get the bug - it did, but also noticed the forward 2 and back 2 didn't work. When back 2 the skipped scene went straight from pushExit to popExit - the pushExit didn't happening because it got cancelled by other state update, so then it tried to do a popExit but that didn't work because the pushExit hadn't reversed yet. So checked if the pushExit had been skipped and cancelled it and went straight to the popExit (reverse pushEnter)