Supported collapsing tabbar on material 3
Turned off liftOnScroll because it [doesn't work with ViewPager](https://github.com/material-components/material-components-android/issues/163). Could get it to work by dynamically setting the liftOnScrollTargetView - but not worth the difficulty for now. For liftOnScroll to work would need to support no_scroll flag on the Toolbar. Even if did that can't configure the colors for the lift on scroll programmatically so not worth the hassle.
Best to turn it off for now to get the collapsing tabbar back