Sketched in idea for back push transition
Fluently navigate from A -> B -> C to A -> D - currently this does a pop enter animation for D. B and D were given same key so it was like a refresh navigation instead of a replace. B and C should exit and D enter - depends on native platform how it shows that though.
Changed the keys to be crumb-stateKey (same as on mobile so scenes are reused if matching state). But then also made sure the top most scene uses the new key - it used to use the previous key. So in the scenario D would have a new key.
This also works for A -> B -> C to A -> D -> C then back 1. The first navigation D has B's key because D isn't on top. Then when going back give D it's own key and then should get B and C exiting and D entering.
Haven't tested it yet but that's the idea.

Also changed Scene to use url to update - copied this over from mobile too. Otherwise B would turn into D when it was exiting. Need B to stay as B and D to be D - so only updating Scene when url matches gives this.

Then changed the native stack to run the replace after the back. Need both to run because it's back to B then replace to D.