Got A -> B to D working on android
It works on ios although it plays the push exit on A when should really play the pop exit. That's ok - it's how it is.
On Android it was trickier. Had to popBackStackImmediate otherwise the replace didn't work - no idea why - tried things like combining the pop backs but didn't work.
Ran the wrong animation by default. It was running the pop Enter on D - makes sense because that's what you'd run on A when popping B off. But now D is treated like a new scene so need to run push enter. Added clause to crumb = oldCrumb on React Native side and returned unmount style for enter instead of crumb style