Checked if back replace
Navigate from A -> B to A then don't want to do back replace
But if go from A -> B to D then do. Used reactViewController to check for now - but planning on switching to a bool on the SceneView to indicate whether it's been added or not. Feels cleaner and don't have to worry about timing - when it's added to window, for example