Turned off the pop animation if back replace
It looked odd to run the pop first and then the replace - the typical scenario is welcome -> login to home. Don't want the welcome scene to come back first and then home to enter and welcome to pop. That's what was happening. Looks better if home enters and welcome exits (ideally pop exit but can't get that on ios)