Removed redundant sort
There was a time when the later keys were longer than earlier. But that was a while back when using the ++ notation for replace. Haven't done that for a while plus there' no need to sort scenes at the same crumb. There's only one scene per crumb in the stack so it doesn't matter which is rendered first