Added stacked bool to indicated if scene added
Was using superview but that's not set for a scene in a crumb. Need to tell apart a back navigation from a back and replace navigation. With the first want to animate because popping to existing scene. With the second don't want to animate because replacing the scene - so don't animate the pop but animate the replace