Used stacked instead of parent for consistency
Can use getParent on android, unlike superview on ios, because it still returns true for scene in crumb. But prefer to match the stacked boolean on ios and also can't guarantee when the parent gets set on android because the fragment stuff is async. Safer and consistent to use own stacked boolean to track once a scene's added to stack