Fixed fluent A -> B -> C to A -> D -> C then back
When swipe or tap back then scene B changed to D. So would get D entering and exiting. That's because the url on B would update to D on the fluent navigation. Changed to update all keys instead of trying to keep the crumbs with the same keys as now.

But then still when swipe back on Android scene B changed into D. Then would get D entering and exiting. That's because the peek navigate would run on B and update state context to D. Added check so only updates context if url matches - same as in getDerivedStateFromProps.

Not sure this is the correct behaviour but it's consistent with other fluent. Maybe should only do push replace when the replaced scene is on top of stack. Here B is crumb when it's replaced so maybe that should silently replace instead of push replace when D appears?

Also but when after swiping back twice to A and then navigating with double taps, instead of navigating 2 scenes it would jump back to A. Set the duration to 1s to give time for the second tap. Turns out it was running the resumeNavigationRef again - guess the rest happened with the second tap. This was probably always a bug but only noticed when testing this.