Changed A -> B -> C to A -> D -> C then back
Don't want this to do a back then push replace because it's too weird. An example that shows how weird is if on stack A -> B -> C and click deep link to D -> E -> C then would get back then push replace when going back to E and then again to D.
Then only time a (back and) push replace should happen is when the change is visible - so when it's the top scene. So, A -> B -> C to A -> D should do back to B then push replace with D. And obviously then one that started it A -> B to A -> C should also do push replace. That's because both cases the scene that's replaced is on top of stack and so visible.
But with A -> B -> C to A -> D -> C the replace isn't visible. So when going back to it the replace has already happened silently so no push replace.
Changed the keys back to how they were but only changed the current key if the scene has changed. So compared state with state of prev crumb in that position and only changed key if they're different. So A -> B -> C to A -> D -> C then back to A -> D - this won't change key because the back shows D hasn't changed. But A -> B -> C to A -> D changes key because D different to B.

Also put peek in Scene back how it was because swipe back should always update. Swipe back will never lead to push replace so it should always update to latest.