Checked if state's are changed
This matches the check in the keys. Checks if, when navigating back, the state is different to the previous state at that crumb - if so then doing a push replace so need different animation because not skipping the back animation in favour of the push replace.

Think it worked before because the back animations aren't used anyway - unless it's a push replace. The back animations already set up when the scene is first pushed onto stack and popped as a crumb