Copied over back replace changes from native
Navigate fluently from A -> B -> C to A -> D then want D to animate with push enter not pop enter because it's new. But navigate fluently from A -> B -> C to A -> B*, where state is the same but data is different, should do pop enter as now. Also A -> B -> C to A -> D -> C then back should do pop enter because only do the push enter when going to that state, not when silently replacing.
Already done these changes on native so copying them over. The key and scene navigation event logic now identical on native and web/mobile