Kept url in sync with navigation
Navigate fluently from A -> B -> C to A -> D then url said B instead of D - so going back and forward in history went back to B instead of D.
So when going back in history to match navigation did a replace history at the end - used the backCrumb to store the end url. Do this whenever there's a backCrumb because that only happens when history is catching up with state context.