Fixed fast double tap back on Android 33,34,35
Tested on 33 and it didn't work. The BackStackEntry count was after the remove, not before like on Android 34. Tried to go back to the old code (before commit a4151794072f9c94a812b6ea0d112143fac3f672) but the isRemoving bug was still there - it was different on Android 33 and 34. Tried to use isVisible in combo with isRemoving and could do it (isVisible ? isVisible : isRemoving) but this didn't work on Android 35. The double tab back went back 3 instead of 2.
So put it back to using BackStackEntry and added <= 33 check to subtract 1 from the entry count. Prefer to have backward check than a forward check - assume it's gonna keep working how it does now in 34 and 35