Started setting insets for edge to edge
The idea is that need to match the height of the navigation bar (and tab bar) to the native height after the insets are applied. For example, with the navigation bar have to add height to match the status bar inset at the top. So going to add the insets to context so that they can be applied correctly on first render for subsequent scenes. For the first scene need something different because can't know the insets at first render - they're calculated dynamically on native side when view attached to window. For that will update the size natively