Updated native size for first scene load
The insets in navigation stack context will ensure the height is correct on first render for all scenes apart from the first. For the first have to get that dynamically on native side and then update the native height to match.
There is a slight flicker because resizing in native is async. So the content jumps down as the navigation bar resizes. Will see if it's any better on the new arch