Corrected the height calculations
Was setting the layout params height when insets change - but this should be calculated automatically by android. Android does the layout params and React Native gets the style height - and these have to match. But removing the params height didn't work - Android  set the height without the top inset. Fixed this by changing measure spec to unspecified. The onMeasure of AppBarLayout only adds the top inset if unspecified.

Another thing. React Native had the wrong width when changing to landscape. The updateNodeSize doesn't have the landscaped width - it just has the new top inset. So changed the width to -1 and that seems to work. React Native sets the correct one