Added top inset to the raw height
Was incorrectly setting the layout params height to include the top inset after the first scene. The layout params height shouldn't include the inset - Android adds the inset as padding and includes it in measured height.
Now the layout params has the raw height - without inset - used it in the React Native size calc and just added the new top inset to it when the insets change