Supported edge-to-edge for collapsing bar
Turned off the measure unspecified on navigation bar when it contains collapsing bar - otherwise the collapsing bar doesn't get a height and android throws error. Could've set layoutparams height on the collapsing bar but prefer to only do unspecified when child is toolbar. Only need this for the coordinatorlayout scenario - on the home scene of twitter - without the unspecified the app bar doesn't have the right height. Think that's because needsCustomLayout is on so it's up to Android to size it rather than React Native.
Also set status bar scrim to transparent so that the navigation bar content shows through - don't want status bar background color with edge-to-edge.
Also set fitsSystemWindows on child view of collapsing so it draws behing the status bar. Again, makes sense for edge-to-edge