Prevented flicker of collapsing toolbar when inset
The first measure and layout after insets applied has to be synchronous else there's a flicker as the toolbar moves into place. This happens with or without collapsing toolbar. But with the collapsing toolbar the problem is trickier because the measure and layout has to happen after the collapsing toolbar's insets - it was happening after navigation bar's insets but before collapsing bar's insets. That's because the navigation bar requests layout and that's sync but that runs from the insets of navigation bar - and that's before collapsing bar's insets because it's parent first. Tried doing the same thing and listening to the inset of the scene but didn't work - because that just set up a boolean before collapsing insets applied and then the layout was requested when navigation bar insets were applied which was too early.
Added a predraw listener in navigation bar and ran a sync measure and layout from there. That's late enough that all insets applied and early enough that there's no flicker