Prevented flicker when scrollable navigation bar
Even though doing a sync measure and layout there was still a flicker when navigating to a scrollable navigation bar - tested navigating from tabs to tabs in twitter sample. Fixed by doing a sync measure and layout of the coordinatorlayout. This also works for the collapsingbar. So now either do sync on parent coordinatorlayout or, if there isn't one, sync on the navigationbar