Prevented toolbar collapsing with material3 theme
On edge-to-edge the pinned toolbar collapses when using material3 theme. This is [an edge-to-edge bug with the collapsingtoolbar](https://github.com/material-components/material-components-android/issues/2608) - it only shows up on material3 theme because it sets forceApplySystemWindowInsetTop to true. So turned this setting off - the setter isn't visible outside of material so get an 'error warning'.
After fixing the scrim color doesn't show up when collapsing with material3 theme - the theme sets a value for scrimVisibilityHeightTrigger which means it doesn't use the height-based calculation anymore. Can't turn this off because not able to set negative value (the default if no theme value provided). So might offer a prop to set this in the future