Made modal sheet edge-to-edge
Took ages to work out how make dialog edge-to-edge. Nothing worked. The key to it all was setting the windowIsFloating flag to false which can only be done from theme resx. After that the standard stuff worked - setting decorFitsSystemWindows to false and the status and navigation bars to transparent. Only did this if there are insets on main activity window because don't want edge-to-edge Dialog if edge-to-edge not enabled.