Got shared elements after updating
In the zoom sample, change color on the details scene then press/swipe back. This didn't work on ios because the shared elements didn't update until the back navigation happened - that was ok for android because the navigation happens on js side first then native. But on ios the navigation back happens native first then js. So updated the shared elements on refresh navigation, too.
This is handy for Android with predictive back, too. But shared elements don't work with predictive back.