Enabled deep linking on android
Tried the LinkActivity but it doesn't work anymore. Calling getLaunchIntentForPackage returns null - think [need to allow querying now](https://stackoverflow.com/a/65140578). But didn't try this because found out that can use the MainActivity. Maybe didn't do this before because used to support activities?! or maybe this didn't used to work?! Not sure. But it works now so will update the docs to use the MainActivity. Not removing the LinkActivity for backward compatibility - either it's working or apps have enabled for querying