Moved into didMoveToWindow for clarity
Thought was gonna have to setPreferredTransition in didSetProps as well as didMoveToWindow. But didMoveToWindow seems to be enough - it's almost always the place where the sheet is first shown because have to wait until it's in window; and because it's removed from tree when dismissed (so close then open adds it back to window). The only time could get didSetProps to do the opening is when non-modal open then navigate and then gesture back. This fires didMoveToWindow before dismissed cleared so didSetProps handles it - but there's no need to setPreferredTransition because it was already set when it was opened