Ensured onDismissed fires before onNavigated
Open non-modal sheet with zoom transition on scene A. Then navigate to scene B and quickly navigate back to A using a custom button. With the new fluid zoom the onNavigated in javascript would fire before the onDismissed from native. That meant the sheet didn't reopen (and wouldn't open again). The timing on zoomed controllers is different to non-zoomed. They stay around longer - take longer to fire the didDisappear events etc. - so other things can happen while they're still open/animating.
Changed to dismiss without animation if the sheet is fluidly zoomed - this fixes the timing. Also it's just better. It looked v odd to animate the closing of the sheet on scene B because it would try to zoom back down to an element that isn't there anymore.
For now have assumed there's only one sheet open but will investigate what happens if there's multiple.