Moved shared elements onto the controller
Shared elements with nested sheets didn't work because the nested sheet couldn't access the scene. The scene wasn't an ancestor of the sheet because it was a descendant of the BottomSheetController not the SceneController. So moved the shared elements onto the Controller so can access them using reactViewController - in scene and first sheet this will return SceneController; in nested sheet it will return BottomSheetController