Dismissed without animation when navigate away
The timing with zoomed controllers is very different - they're much slower to complete/dismiss then traditional controllers. So with non-modal open and  navigate away and back quickly the onNavigated fires before the onDismiss and then the sheet doesn't reopen. So turned off the animation for zoomed sheets - it was better without because otherwise the sheet zooms back to nothing because it's different scene. But this it's too tricky to do this when there are multiple sheets open - checking if any of them are zoomed is hard - and there's no gain. Better to disable the animation completely because it's consistent. Plus avoids potential timing issues with even non-zoomed sheets if the back happens super fast or the dismiss is slow