Sketched in rsc plan
Already tested out the rough idea in the parcel rsc repo, https://github.com/parcel-bundler/rsc-examples. The idea is that the SceneView will work server side - and it will wrap the child content in a client component, SceneRSCView. That way only the active content will be included in the bundle - it works as long as the content is React lazy imported (or dynamic import).
Then the SceneRSCView can refresh just the content instead of the whole page. So it's combo of the client and server examples from parcel rsc repo. The server example for the initial ssr load and then the client example to dynamically update just the changed content after that.
Other part of plan is useNavigationEvent will work both client and server. It will return the navigation context. On the server this uses the react cache from react 19. Only way could get this working is to mutate the cached value