Created rsc sample
Already tested out the rough idea in the parcel rsc repo, https://github.com/parcel-bundler/rsc-examples. Copied over the server and client files and all the setup from that repo. Already checked that can SSR the App for first request and it only sends down the bundle for the active scene view. Then can update the SceneView by sending updates (matching the client sample from the repo) for just that RSC.
It's not working yet because haven't built separate navigation-react package for react-server environment - it's complaining that there's a createContext call when try and start the server.