Fetched async data!
It wouldn't work because think there's a bug with parcel. It's not respecting the 'use client' directive inside navigation-react. For example, rendering NavigationLink inside People component should create 2 People bundles - one for server and one for client because NavigationLink is client component. But it only ever generated one. Wrapping it in a 'use client' component inside the App fixed it.
Also had to do the same for NavigationHandler. Had to pass the url down so could navigateLink on the 'client' - can't call stateNavigator.start because the client code runs on the server when doing SSR (want to avoid a separate export build for node client/default) and can't get hold of window url in this case. So passed down the url so it works on server and also on client - where the url prop is serialized so code runs the same on the client