Made SceneRSCView client component and exported
It needs to be exported from the client bundle even though it's rendered from the server one - because it's a client component so it's actually the client bundle that renders it.
Running into the same parcel bundler problem where parcel isn't respecting the 'use client' directive in navigation react. Going to try and create a minimal repro so can raise against parcel