Got parcel to 'use client' in navigation react
Was going to raise a bug against parcel so started trying to create a minimal repo. Had real trouble trying to create the 'npm module' to test the 'use client'. Not sure what happened but switched from parcel watch to parcel build - but it didn't like the module.exports syntax in npm module. Kept getting parcel error about no file or directory and then some file with runtime in the name?! So tried esm exports instead and parcel build like it!! In fact it even then did respect the use client!! The parcel watch just didn't work at all - it didn't generate all the files that parcel build did - and then get the error that started this investigation (expected a string (for built-in components) or a class/function) - so actually think the problem is with parcel watch!!
So switched back to navigation-react and built with esm export instead (target es2022 and module nodenext) and used parcel build and it worked! Don't need the Links 'use client' wrapper anymore!!
Not sure can release like this but at least can continue for now