Rendered SceneRSCView even if inactive
Has to render otherwise it can't update on the client. The rsc fetch isn't working - well it's being called infinitely, not sure why.
Also realised the active (and refetch/update) prop has a problem - the values have to be serializable for rsc so there's can't pass a function?! will have to rethink