Implemented name filter using deferred value!!
The problem is that the filter input doesn't update until the search results are returned. If they take 2s to come back then it's not possible to type in the filter input.
So useDeferredValue for the stateContext. It's exactly [the scenario from the React docs](https://react.dev/reference/react/useDeferredValue). This allows the new stateContext to be committed right away but still display the old search results. So the filter input updates instantly and the search results fetch happens in the background. It's committed when it returns.
Only put this in for refresh navigation because that's where stay on the current page. Don't want this if not refresh because want to commit the new state context with any fetch results at the same time. Otherwise, for example, going back from person to people then the back link would disable because it's instantly on people now but would still be showing person because that's the results for the old context. Don't want that