Moved from useDeferredValue to startTransition
Using deferred isn't good. Put it in originally because it got round the problem of typing in filter text box - without deferred it's the input doesn't update till the async completes which is too slow for typing. But it doesn't work for other reasons, for example, typing in filter then the paging links update immediately so could type in 'saefsd' in filter and then press Next page which clearly doesn't exist - the Next page is enabled because it's using the old/deferred state context. Really have to wait for the results to come back before updating. This is what the default async does so removing deferred fixes this. But then get the Suspense boundary triggered with each search. That's where startTransition comes in - it prevents the previous search results from being replaced with suspense boundary because React won't hide previously shown content. But can't ask developers to add startTransition to RefreshLinks because can't pass functions props from server components to client components - props must be serializable. So would have to make the pager a client component which is bad. So changed NavigationReact to use startTransition by default - this is the React recommendation for RSC-supporting routers but now understand why