Implemented nested scene views
There's no need to lazy load the nested scene view. To lazy load would need the SceneView active prop to work on data not just state keys - would need active to return false if show isn't true. Can't do that yet and don't think it's necessary - having separate bundle per scene is enough. Don't think there's any gain having separate bundles within a scene. The important thing is to independently fetch the nested scene view so that the whole scene doesn't reload. Currently it is all reloading - and that does work - but need to implement the fetch data dependency prop on SceneView so only refetch when that data's changed.
But also currently it's fetching the person scene and the friends scene separately. Need to make the SceneView aware of ancestor SceneViews so that they don't fetch if an ancestor is fetching. Friends will get fetched by the person fetch in this example