Implemented conditional refetch with data deps
Added a dataDeps array prop to SceneView that holds the keys of the navigation data that determine when the SceneView needs to update. Tested on person page by adding an extra link that changes person too. It wasn't working because when the Person scene doesn't update there was no sceneView in the cache to show. So stored the latest fetched scene in a useRef. Then can try to show the new fetched scene if it's there and then try to show the one from the ref if it isn't there - and as last resort show the children if never fetched any scene update at all. It seems to work?!
Should be able to get nested SceneViews working too now this cache and fallback strategy is in place. Only tested basic scenario so not sure if the strategy is correct.