Renamed fetched and rendered SceneView for clarity
The fetchedSceneView one is from a createFromFetch rsc call. The renderedSceneView is set to the fetchedSceneView once the async changes are committed. Then render will always render the fetchedSceneView if there is one - then the renderedSceneView if there isn't - and fallback to the children if there's never been a fetched scene