Implemented nested scene views
Assumed the context changes always propagate down through the tree - so that the parent gets the context changes before the child. Seems that's how it would have to work but not sure?! Assuming that makes it simpler to check from a descendant rsc scene view whether an ancestor is refetching because it's data dependencies have changed. Can keep asking the ancestor provider whether they've changed.
So if the ancestor is fetching then the descendant mustn't fetch - there's no need because it's data will be provided by the ancestor fetch. But the tricky part is in that case it must go back to rendering children instead of the fetched scenes. So checked ancestor fetch to ignore the fetchedSceneView and the renderedSceneView. Also need to clear the renderedSceneView when React commits the tree so that next time round children are used instead of the last fetched scene view.