Sketched in rsc navigation idea
Checked on next and the redirect is a throw error that's caught on the client - guessing with an error boundary - and then client side navigation. The error is streamed in the rsc response. The next redirect in the stream looks like `d:E{"digest":"NEXT_REDIRECT;replace;/;307;","message":"NEXT_REDIRECT"...`.
Tried the same by throwing an error from Person scene. The error is streamed but can't seem to populate the digest - not sure if this is a parcel bug (guessing so?). Anyway, can use the message instead. So wrapped the RSC scene view in a navigation error boundary and navigated the link in the error message and it works.
Need to throw the error from the server navigateLink and also rethrow client side if not a navigation link 'error'