Handle the server navigation on the client
The error boundary checks the message to see if it's a server navigation. If it is then it calls navigateLink.
Tried to navigate from within componentDidCatch but the timing was wrong - the componentDidCatch fires before React commits async rendering. So if navigate from there the state hasn't committed and the navigation isn't completed - the state context isn't updated, history not added etc. This only happens after React commits the state. So moved the actual navigation into a useEffect to make sure it happens after the state changes are committed.
So in test example, called navigate inside Person so the url quickly updates to person before the 'error' boundary navigation runs. Think this is correct because the Person scene might have suspense boundary so the url should update to person. Changed the default historyAction to replace instead of add because don't think want to go back to the error'ed person page. Typical scenario is redirecting because not logged in anymore or don't have permission - something like that?

Defaulted the historyAction to replace on the server StateNavigator because