Supported parcel's hrml reload
When hmr reload happens the client has the old code and the server has the new code. Need to rerender the App on the server so that latest code comes down. Need to do with rsc fetch so client state is retained.
Copied the idea from the hotloader sample in repo. On the client, turn the current url into crumb/state keys and data and send to server. The server turns that back into the new url using latest state navigator. Then render the App with that url - same as the normal get request.
It works ok when on people page but not when on person page. Not sure why, but when updating the stateNavigator, by changing route for example, this does the rsc fetch on people page - but on person page it doesn't, it triggers a full page reload and throws error in get request because the url isn't valid anymore. Will raise with parcel