Ensured renders when history not cached
Navigate from  anywebsite.com --> People --> Person then history back twice to anywebsite then history forward twice to Person. This didn't render the Show Friends link. This is situation where browser history but history isn't cached because unloaded the example when went back to anywebsite.com. Friends was incorrectly thinking that history was cached because useLayoutEffect fired (testing with Suspense boundary in Person) for People and Person when Suspense returned. So friends went to cache to render but there was nothing there.
Fixed by checking for the sceneViewKey in history cache not just the presence of the url in history cache