Supported reloading the scene using refresh
After updating data need to reload the page - not using server actions, so need to support programmatic reload following save.
Took refresh with current data as the reload scenario
{data} = useNavigationEvent();
refresh({...data}) // reload
Don't want to trigger an actual navigation because
- no navigation has happened
- don't want a new state context with different old state and data
- might get blocked if user has onNavigating handler that returns false
- won't reload because there's no changed data
Instead of navigating, mutated the navigation event context! So keep the exact same stateContext but just {...navigationEvent} to build new one.
Needed to add ignoreCache boolean to the navigation event context so that the scene view doesn't use the history cache. Also used this to turn off the oldUrl check because need it to go back to the server to reload all data.
Also had to change the setState(() => navigationEvent) to setState(navigationEvent) is AsyncStateNavigator - otherwise ended up with infinite data fetch. React would keep calling the state setter, generating a new navigationEvent each time and then meant there was no fetch in the cache because the cache is keyed off the navigaitonEvent now. This is a good change generally because probably is currently over rendering by returning new navigationEvent each time