Reloaded after history navigation to prevent stale
History navigation loads from cache - it has to because
- browser expects instant content to restore scroll position
- react does a synch update (to restore scroll) and this would reveal suspense loading if fetch from server
But leaving on cached means the data could be stale, so triggered an immediate reload after history navigation. Used same technique already in place for reloading - setState with {...navigationEvent} to ensure state updates and rerenders and also to keep the existing stateContext - don't want to change any state context because it's not a navigation. Also pass ignoreCache to ensure fresh data fetched.
Also reused the same asyncStateNavigator when reloading - there's no need for a new one because the state context is the same