Cleared rsc cache when React commits
Not sure if this is correct but it seems right. Once React commits then the rsc cache is irrelevant for everything except the current scene (navigationEvent). All in flight background async will be abandoned once React commits - if click 3 links quickly, for example, then React will commit the last and abandon the first 2. So when React commits only need 3 in the cache