Cleared history cache
If go from A --> B then browser back and A --> C then can clear B from the history cache. Kept track of a count when navigating and stored this in browser history state. Calculate the next count by adding 1 to the current count. Don't increment when browsing history. Stored the count against the history cache, too. Then when navigate can clear out any cache items with count >= history state count. So in example, A has count 1 and B 2 then C also has 2 and so can clear B when navigate to C.
Change from useLayoutEffect to useEffect so that the url is updated. Think useEffect is better anyway because doesn't need to be sync