Ensured sceneKey doesn't change within scene
In the example go from people to person with show true to expand friends section and wrap friends in suspense with delayed loading. Then the sceneCount incremented and was stored in browser state by the time friends came in and then got incremented again. So person and friends had different sceneCounts.
Changed to reuse the sceneCount if it's there and only incremented if it's not - so the first time.
Need to handle historyAction none and replace because can't delete in these cases - can still browser forward after them. But also need to prevent history cache if historyAction is none. So will need to add historyAction to stateContext and then handle