Moved code into src to match upcoming webpack
Been working on rsc-webpack and that has code in the src folder - way to hard to move folders with the webpack config so going with src for consistency. Plus it matches the original parcel example anyway and prefer to stick with that.
No idea why but the parcel-cache folder got moved to the root navigation - so explicitly added cacher-dir cli option to locate it inside the sample folder