Supported href not matching url (e.g., appPath)
It could be that the server is hosted at '/app/...' rather than at the root, with new HTML5HistoryManager('app') (and a router prefix on the server app.use('/app', router)). In that case mustn't pass the url down, need to pass the href down or else the server will 404. This could happen for other scenarios where href doesn't match url. So sent the state context url in the body instead and used getHref to get the request url/href.
Interestingly don't have to pass the appPath to the StateNavigators created in the server - only the one created in RootProvider!! That's because Navigation Links are client components and so get their stateNavigator from the RootProvider on React.Context!!!