Supported appPath in webpack rsc
Already done for parcel so copying over to webpack. It's trickier because it's webpack, plus 2 express app's. But essentially to setup change the publicPath to '/app/' in webpack.config and prefix '/app/' in main bundle. Then the region are similar to the parcel ones with a router prefix app.use('/app', router)